test(SyncMessageService): Remove operators limiting number of events observed
We want to make sure that, for each input, a determined number of outputs are generated.
By limiting the number of of events observed, we were only testing that at least
this number of outputs were generated.
By removing these operators, we are making sure that no additional and unexpected ouputs are emitted.